working carousel as separate component
